refactor: move workspace service
